Move BaseWorkerContext testing to BaseWorkerContextTests
